# P-384/521 runtime check for s2n-bignum on aarch64